oracle_db_monitor/requirements.txt
Flask==2.0.1
python-oracledb==1.3.1
gunicorn==20.1.0
python-dotenv==0.19.0
